Rename from precommit_nbconvert_rename to nb_prep
