+++ README.md
@@ -1,12 +1,12 @@
-[![Unit tests](https://github.com/allianz-direct/precommit_nbconvert_rename/actions/workflows/unit_tests.yml/badge.svg)](https://github.com/allianz-direct/precommit_nbconvert_rename/actions/workflows/unit_tests.yml)
-![PyPI - Python Version](https://img.shields.io/pypi/pyversions/precommit-nbconvert-rename)
-![PyPI](https://img.shields.io/pypi/v/precommit-nbconvert-rename)
-![PyPI - Downloads](https://img.shields.io/pypi/dm/precommit-nbconvert-rename)
-![GitHub contributors](https://img.shields.io/github/contributors/timvink/precommit-nbconvert-rename)
-![PyPI - License](https://img.shields.io/pypi/l/precommit-nbconvert-rename)
+[![Unit tests](https://github.com/allianz-direct/nb_prep/actions/workflows/unit_tests.yml/badge.svg)](https://github.com/allianz-direct/nb_prep/actions/workflows/unit_tests.yml)
+![PyPI - Python Version](https://img.shields.io/pypi/pyversions/nb-prep)
+![PyPI](https://img.shields.io/pypi/v/nb-prep)
+![PyPI - Downloads](https://img.shields.io/pypi/dm/nb-prep)
+![GitHub contributors](https://img.shields.io/github/contributors/timvink/nb-prep)
+![PyPI - License](https://img.shields.io/pypi/l/nb-prep)
 
 
-# precommit_nbconvert_rename
+# nb_prep
 
 Use `nbconvert` and `nbstripout` together as precommit hooks.
 
@@ -23,7 +23,7 @@ You use [jupyter notebooks](https://jupyter.org/) and:
 
 Forget to run `nbconvert`, or use them in the wrong order (`nbstripout` before `nbconvert`) and you will have to re-run your notebooks before you can output HTML, which can be annoying when they are long-running. Especially when you use `nbstripout` as a [pre-commit](https://pre-commit.com/) hook, this can happen quite often.
 
-`precommit_nbconvert_rename` can help to automatically process notebooks and (optionally) store versioned output in an in output directory. The CLI command `nb_convert_strip` takes a list of directories and/or files to find and convert notebooks. For each notebook:
+`nb_prep` can help to automatically process notebooks and (optionally) store versioned output in an in output directory. The CLI command `nb_convert_strip` takes a list of directories and/or files to find and convert notebooks. For each notebook:
 
 - `nbconvert` is used to create an `<filename>.html` export
 - Date prefix is added `YYYYMMDD_<filename>.html` (can be turned off)
@@ -41,7 +41,7 @@ You can setup this entire workflow once as [pre-commit](https://pre-commit.com/)
 ## Installation
 
 ```bash
-pip install precommit_nbconvert_rename
+pip install nb_prep
 ```
 
 ## Usage [TODO, update]
@@ -54,7 +54,7 @@ repos:
 -   repo: local
     hooks:
     -   id: nbconvert_rename_precommit
-        name: precommit_nbconvert_rename (pre-commit; run nbconvert)
+        name: nb_prep (pre-commit; run nbconvert)
         description: 'Converts to .ipynb to .html and adds date prefix and hash placeholder.'
         entry: nbconvert_rename
         language: python
@@ -62,7 +62,7 @@ repos:
         types: [jupyter]
         stages: [commit]
     -   id: nbconvert_rename_postcommit
-        name: precommit_nbconvert_rename (post-commit; replace commithash in .html filenames)
+        name: nb_prep (post-commit; replace commithash in .html filenames)
         description: 'Replaces NBCONVERT_RENAME_COMMITHASH_PLACEHOLDER with commit hash in any .html filenames.'
         entry: rename_commithash
         types: [html]    
@@ -71,6 +71,16 @@ repos:
         always_run: true
         stages: [post-commit]
 ```
+
+```yaml
+repos:
+-   repo: https://github.com/psf/black
+    rev: stable
+    hooks:
+    - id: black
+```
+
+
 
 You need to install the pre-commit and the post-commit hooks separately:
 
@@ -84,11 +94,11 @@ When you commit a notebook, you might see something like:
 ```shell
 git add notebook.ipynb
 git commit -m "Add notebook"
-# precommit_nbconvert_rename (pre-commit; run nbconvert)............................Passed
+# nb_prep (pre-commit; run nbconvert)............................Passed
 # nbstripout........................................................................Failed
 # - hook id: nbstripout
 # - files were modified by this hook
-# precommit_nbconvert_rename (post-commit; replace commithash in .html filenames)...Passed
+# nb_prep (post-commit; replace commithash in .html filenames)...Passed
 ```
 
 `nbstripout` has overwritten `notebook.ipynb` and `nbconvert-rename` has created a file named something like `20211026_notebook_NBCONVERT_RENAME_COMMITHASH_PLACEHOLDER.html`.
@@ -97,9 +107,9 @@ Make sure to avoid committing HTML files by adding `.html` added to your `.gitig
 ```shell
 git add notebook.ipynb
 git commit -m "Add notebook"
-# precommit_nbconvert_rename (pre-commit; run nbconvert)............................Passed
+# nb_prep (pre-commit; run nbconvert)............................Passed
 # nbstripout........................................................................Passed
-# precommit_nbconvert_rename (post-commit; replace commithash in .html filenames)...Passed
+# nb_prep (post-commit; replace commithash in .html filenames)...Passed
 ```
 
 Now, you've committed a clean, stripped version of `notebook.ipynb` and you have a local snapshot of your notebook named something like `20211026_notebook_eac9e43.html`.
