+++ nb_prep/_utils.py
@@ -3,6 +3,10 @@ import subprocess
 
 
 class GitError(Exception):
+    """
+    Exception with Git.
+    """
+
     pass
 
 
@@ -25,9 +29,7 @@ def git_version():
         env["LANGUAGE"] = "C"
         env["LANG"] = "C"
         env["LC_ALL"] = "C"
-        sp = subprocess.Popen(
-            cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=env
-        )
+        sp = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=env)
         out, err = sp.communicate()
         if sp.returncode != 0:
             raise GitError(err.strip().decode("ascii"))
