+++ nb_prep/nb_convert_strip.py
