+++ tests/data/pre-commit-test-config.yaml
@@ -2,7 +2,7 @@ repos:
 -   repo: local
     hooks:
     -   id: nbconvert_rename_precommit
-        name: precommit_nbconvert_rename (pre-commit; run nbconvert)
+        name: nb_prep (pre-commit; run nbconvert)
         description: 'Converts to .ipynb to .html and adds date prefix and hash placeholder.'
         entry: nbconvert_rename
         language: python
@@ -10,7 +10,7 @@ repos:
         types: [jupyter]
         stages: [commit]
     -   id: nbconvert_rename_postcommit
-        name: precommit_nbconvert_rename (post-commit; replace commithash in .html filenames)
+        name: nb_prep (post-commit; replace commithash in .html filenames)
         description: 'Replaces NBCONVERT_RENAME_COMMITHASH_PLACEHOLDER with commit hash in any .html filenames.'
         entry: rename_commithash
         types: [html]
