+++ tests/test_convert_notebook.py
@@ -3,8 +3,8 @@ import shutil
 from freezegun import freeze_time
 from pathlib import Path
 
-from precommit_nbconvert_rename.nb_convert_strip import convert_notebook
-from precommit_nbconvert_rename.files import working_directory
+from nb_prep.nb_convert_strip import convert_notebook
+from nb_prep.files import working_directory
 
 
 @freeze_time("2012-01-14")
